Rework content-script with reducer pattern and signals
