+++ app/package.json
@@ -58,6 +58,7 @@
     "@aside/core": "0.7.4",
     "@aside/chrome-ui": "0.7.5",
     "react": "18.2.0",
-    "@remote-ui/react": "5.0.2"
+    "@remote-ui/react": "5.0.2",
+    "@preact/signals-core": "^1.1.5"
   }
 }
